fix: k8s refresh token

+++ script/service/gcp-refresh-token.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-rm /root/.kube/config
+# Check if the kubeconfig file exists and remove it if it does
+if [ -f /root/.kube/config ]; then
+  rm /root/.kube/config
+fi
+
+# Ensure the script is executed from the /app directory
+cd /app
 
 # Authenticate with the service account
 gcloud auth activate-service-account --key-file=service-account/gcp.json
